add some verification

+++ shop/sources/template/Default/modules/modules_account_customers/ac_account_customers_edit.php
@@ -32,7 +32,7 @@
       $this->description = CLICSHOPPING::getDef('module_account_customers_edit_description');
 
       if (defined('MODULE_ACCOUNT_CUSTOMERS_EDIT_TITLE_STATUS')) {
-        $this->sort_order = MODULE_ACCOUNT_CUSTOMERS_EDIT_TITLE_SORT_ORDER;
+        $this->sort_order = (int)MODULE_ACCOUNT_CUSTOMERS_EDIT_TITLE_SORT_ORDER;
         $this->enabled = (MODULE_ACCOUNT_CUSTOMERS_EDIT_TITLE_STATUS == 'True');
       }
     }
